+added function AbsGradientSaturatedSum;

+++ trunk/src/Simd/SimdAbsSecondDerivativeHistogram.cpp
@@ -3,20 +3,20 @@
 *
 * Copyright (c) 2011-2013 Yermalayeu Ihar.
 *
-* Permission is hereby granted, free of charge, to any person obtaining a copy 
+* Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
-* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell 
-* copies of the Software, and to permit persons to whom the Software is 
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
-* The above copyright notice and this permission notice shall be included in 
+* The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
-* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
-* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
-* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE 
-* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER 
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
@@ -57,14 +57,14 @@ namespace Simd
 					const int sdX = AbsSecondDerivative(src + col, step);
 					const int sdY = AbsSecondDerivative(src + col, rowStep);
 					const int sd = MaxU8(sdY, sdX);
-					++histogram[sd];					
+					++histogram[sd];
 				}
 				src += stride;
 			}
 		}
 	}
 
-#ifdef SIMD_SSE2_ENABLE    
+#ifdef SIMD_SSE2_ENABLE
 	namespace Sse2
 	{
 		namespace
@@ -125,7 +125,7 @@ namespace Simd
 					AbsSecondDerivative<false>(src, step, rowStep, buffer.p);
 				for(ptrdiff_t col = bodyStart; col < bodyEnd; col += A)
 					AbsSecondDerivative<align>(src + col, step, rowStep, buffer.p + col);
-				if(width != bodyEnd)
+				if(width != (size_t)bodyEnd)
 					AbsSecondDerivative<false>(src + width - A, step, rowStep, buffer.p + width - A);
 
 				for(size_t i = 0; i < width; ++i)
